# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_transport_rest_handler.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: Invalid input
        type: string
      message:
        example: The provided data is invalid
        type: string
    type: object
  internal_transport_rest_handler.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  sarc-ng_internal_domain_reservation.Reservation:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      endTime:
        type: string
      id:
        type: integer
      purpose:
        type: string
      resourceID:
        type: integer
      startTime:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
...
  sarc-ng_internal_service_reservation.CreateReservationDTO:
    properties:
      endTime:
        type: string
      resourceId:
        type: integer
      startTime:
        type: string
      status:
        type: string
      userId:
        type: integer
    required:
    - endTime
    - resourceId
    - startTime
    - userId
    type: object
  sarc-ng_internal_service_reservation.UpdateReservationDTO:
    properties:
      endTime:
        type: string
      id:
        type: integer
      resourceId:
        type: integer
      startTime:
        type: string
      status:
        type: string
      userId:
        type: integer
    required:
    - endTime
    - id
    - resourceId
    - startTime
    - userId
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: SARC Next Generation - Resource Reservation and Management API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SARC-NG API
  version: "1.0"
paths:
...
  /reservations:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all reservations in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of reservations
          schema:
            items:
              $ref: '#/definitions/sarc-ng_internal_domain_reservation.Reservation'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Get all reservations
      tags:
      - reservations
    post:
      consumes:
      - application/json
      description: Create a new reservation with resource, user, and time information
      parameters:
      - description: Reservation creation data
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/sarc-ng_internal_service_reservation.CreateReservationDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created reservation
          schema:
            $ref: '#/definitions/sarc-ng_internal_domain_reservation.Reservation'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Create a new reservation
      tags:
      - reservations
  /reservations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a reservation by its ID
      parameters:
      - description: Reservation ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservation deleted successfully
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.SuccessResponse'
        "400":
          description: Invalid reservation ID
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "404":
          description: Reservation not found
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Delete a reservation
      tags:
      - reservations
    get:
      consumes:
      - application/json
      description: Retrieve a specific reservation by its unique identifier
      parameters:
      - description: Reservation ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservation details
          schema:
            $ref: '#/definitions/sarc-ng_internal_domain_reservation.Reservation'
        "400":
          description: Invalid reservation ID
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "404":
          description: Reservation not found
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Get reservation by ID
      tags:
      - reservations
    put:
      consumes:
      - application/json
      description: Update an existing reservation's resource, user, time, and status
        information by ID
      parameters:
      - description: Reservation ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Reservation update data
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/sarc-ng_internal_service_reservation.UpdateReservationDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Updated reservation
          schema:
            $ref: '#/definitions/sarc-ng_internal_domain_reservation.Reservation'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "404":
          description: Reservation not found
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Update an existing reservation
      tags:
      - reservations
...
schemes:
- http
- https
swagger: "2.0"